chore(fill): duplicate `FixtureAuthorizationTuple`'s `v` field as `yParity`
