+++ src/ethereum_test_fixtures/blockchain.py
@@ -34,7 +34,6 @@ from ethereum_test_base_types import (
 from ethereum_test_exceptions import EngineAPIError, ExceptionInstanceOrList
 from ethereum_test_forks import Fork, Paris
 from ethereum_test_types.types import (
-    AuthorizationTupleGeneric,
     Transaction,
     TransactionFixtureConverter,
     TransactionGeneric,
@@ -43,7 +42,7 @@ from ethereum_test_types.types import (
 )
 
 from .base import BaseFixture
-from .common import FixtureBlobSchedule
+from .common import FixtureAuthorizationTuple, FixtureBlobSchedule
 
 
 class HeaderForkRequirement(str):
@@ -316,19 +315,6 @@ class FixtureEngineNewPayload(CamelModel):
         )
 
         return new_payload
-
-
-class FixtureAuthorizationTuple(AuthorizationTupleGeneric[ZeroPaddedHexNumber]):
-    """Authorization tuple for fixture transactions."""
-
-    signer: Address | None = None
-
-    @classmethod
-    def from_authorization_tuple(
-        cls, auth_tuple: AuthorizationTupleGeneric
-    ) -> "FixtureAuthorizationTuple":
-        """Return FixtureAuthorizationTuple from an AuthorizationTuple."""
-        return cls(**auth_tuple.model_dump())
 
 
 class FixtureTransaction(TransactionFixtureConverter, TransactionGeneric[ZeroPaddedHexNumber]):
